test 5 works

+++ values.cc
@@ -12,38 +12,36 @@ using namespace std;
 #include "values.h"
 #include "listing.h"
 
-int evaluateReduction(Operators operator_, int head, int tail)
+double evaluateReduction(Operators operator_, double head, double tail)
 {
 	if (operator_ == ADD)
 		return head + tail;
 	return head * tail;
 }
 
-
-int evaluateRelational(int left, Operators operator_, int right)
+double evaluateRelational(double left, Operators operator_, double right)
 {
-	int result;
+	double result;
 	switch (operator_)
 	{
-		case LESS:
-			result = left < right;
-			break;
+	case LESS:
+		result = left < right;
+		break;
 	}
 	return result;
 }
 
-int evaluateArithmetic(int left, Operators operator_, int right)
+double evaluateArithmetic(double left, Operators operator_, double right)
 {
-	int result;
+	double result;
 	switch (operator_)
 	{
-		case ADD:
-			result = left + right;
-			break;
-		case MULTIPLY:
-			result = left * right;
-			break;
+	case ADD:
+		result = left + right;
+		break;
+	case MULTIPLY:
+		result = left * right;
+		break;
 	}
 	return result;
 }
-
